% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_optimizer.R
\name{ci_optimizer}
\alias{ci_optimizer}
\title{Optmization algorithm based on fmincon}
\usage{
ci_optimizer(x)
}
\arguments{
\item{x}{A Dataframe}
}
\value{
A data frame
}
\description{
Optmization algorithm based on fmincon
}
\examples{
ci_optimizer(x)
}
