% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{scaling}
\alias{scaling}
\title{Normalization and standardization techniques}
\usage{
scaling(x, method = "min-max")
}
\arguments{
\item{x}{A Dataframe}

\item{method}{Standardization or normalization technique. So far "min-max" and "standardization" are available}
}
\value{
A data frame
}
\description{
Normalization and standardization techniques
}
\examples{
x <- data.frame(rnorm(20),rnorm(20),rnorm(20),rnorm(20))
scaling(x,method = "min-max")
}
