% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_linear.R
\name{si_linear}
\alias{si_linear}
\title{Calculate Si using linear method}
\usage{
si_linear(x, avg_type = "simple")
}
\arguments{
\item{x}{A Dataframe}

\item{avg_type}{Choosing average type. So far "simple", "geometric" and "harmonic" average are availableç}
}
\value{
A data frame
}
\description{
Calculate Si using linear method
}
\examples{
si_linear(x,avg_type = "simple")
}
