% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_linear_exc.R
\name{si_linear_exc}
\alias{si_linear_exc}
\title{Calculate Si using linear method by excluding Xi}
\usage{
si_linear_exc(x, avg_type = "simple")
}
\arguments{
\item{x}{A Dataframe}

\item{avg_type}{Choosing average type. So far "simple", "geometric" and "harmonic" average are availableç}
}
\value{
A data frame
}
\description{
Calculate Si using linear method by excluding Xi at each iteration while calculating Si
}
\examples{
si_linear_exc(x,avg_type = "simple")
}
