% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_linear_exc_vif.R
\name{si_linear_exc_vif}
\alias{si_linear_exc_vif}
\title{Calculate Si using linear method by excluding Xi using VIF}
\usage{
si_linear_exc_vif(x, avg_type = "simple", vif_threshold = 4.5)
}
\arguments{
\item{x}{A Dataframe}

\item{avg_type}{Threshold for VIF. Based on this threshold variables from input data (x) are excluded for the calculations.}
}
\value{
A data frame
}
\description{
Calculate Si using linear method by excluding Xi using VIF
}
\examples{
si_linear_exc_vif(x,avg_type = "simple", vif_threshold = 4.5)
}
