% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_average.R
\name{calc_average}
\alias{calc_average}
\title{Calculate averages}
\usage{
calc_average(x, avg_type = "simple")
}
\arguments{
\item{x}{A Dataframe}

\item{avg_type}{Choosing average type. So far "simple", "geometric" and "harmonic" average are availableç}
}
\value{
A data frame
}
\description{
Calculate different types of averages
}
\examples{
calc_average(x,avg_type = "simple")
}
